trabajoFinal/programa/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
filelock==3.15.4
pillow==10.4.0
platformdirs==4.2.2
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
